FlashCover ではなく PageRoom に音を鳴らす機能を持たせた
