+++ src/FlashCover.svelte
@@ -1,12 +1,10 @@
 <script lang="ts">
   import clsx from "clsx"
-  import { filter, map, Observable, withLatestFrom } from "rxjs"
-  import { onMount } from "svelte"
+  import { map, Observable, withLatestFrom } from "rxjs"
   import { mapToCurrentDuration } from "./mapToCurrentDuration"
   import { notifyFirstZero } from "./notifyFirstZero"
   import { now } from "./now"
   import type { TimerState } from "./timerReducer"
-  import { useAudio } from "./useAudio"
   import { useConfig } from "./useConfig"
   import { interval } from "./util/interval"
 
@@ -22,26 +20,6 @@
     withLatestFrom(config$),
     map(([_, config]) => config.flash === "on" && _)
   )
-
-  $: sounding$ = timerState$.pipe(
-    mapToCurrentDuration(interval("worker", 100), now),
-    notifyFirstZero(),
-    withLatestFrom(config$),
-    map(([_, config]) => config.sound === "on" && _)
-  )
-
-  const audio = useAudio()
-
-  onMount(() => {
-    const sub = sounding$.pipe(filter((_) => _ === true)).subscribe(() => {
-      console.debug("audio.play()")
-      audio.play()
-    })
-
-    return () => {
-      sub.unsubscribe()
-    }
-  })
 </script>
 
 <div
