+++ src/PageRoom.svelte
@@ -1,13 +1,27 @@
 <script lang="ts">
-  import { firstValueFrom, map, merge, partition, type Observable } from "rxjs"
+  import {
+    filter,
+    firstValueFrom,
+    map,
+    merge,
+    partition,
+    withLatestFrom,
+    type Observable,
+  } from "rxjs"
   import { onMount } from "svelte"
   import FlashCover from "./FlashCover.svelte"
+  import { mapToCurrentDuration } from "./mapToCurrentDuration"
+  import { notifyFirstZero } from "./notifyFirstZero"
+  import { now } from "./now"
   import { isRoom, type InvalidDoc, type Room } from "./schema/roomSchema"
   import { setTitleAsTimeViewer } from "./setTitleAsTimeViewer"
   import Timer from "./Timer.svelte"
+  import { useAudio } from "./useAudio"
+  import { useConfig } from "./useConfig"
   import { useRoom } from "./useRoom"
   import { useSetup } from "./useSetup"
   import { useTimerState } from "./useTimerState"
+  import { interval } from "./util/interval"
 
   export let roomId: Room["id"]
 
@@ -16,6 +30,26 @@
   $: timerState$ = useTimerState(roomId)
 
   onMount(() => setTitleAsTimeViewer(timerState$))
+
+  const config$ = useConfig()
+  $: sounding$ = timerState$.pipe(
+    mapToCurrentDuration(interval("worker", 100), now),
+    notifyFirstZero(),
+    withLatestFrom(config$),
+    map(([_, config]) => config.sound === "on" && _)
+  )
+
+  const audio = useAudio()
+  onMount(() => {
+    const sub = sounding$.pipe(filter((_) => _ === true)).subscribe(() => {
+      console.debug("audio.play()")
+      audio.play()
+    })
+
+    return () => {
+      sub.unsubscribe()
+    }
+  })
 
   function roomOrInvalid(_room$: Observable<Room | InvalidDoc>) {
     const [room$, _invalid$] = partition(_room$, isRoom)
